moved checks to components
