+++ src/components/Checks.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { styled } from "./stitches.config";
+import { styled } from "../stitches.config";
 
 const Wrapper = styled("div", {
   display: "flex",
